+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Fixed
+- Added sanitize function in `string-utils` and using in `history` label in `Autocomplete`
+
 ## [2.2.0] - 2020-12-02
 
 ### Added